Exempt AutoValue-generated equals methods from EqualsIncompatibleType.
AutoValue is often used as a "bag of things", with no real intention to use it for equals/hashCode. This happens a lot with, say, mutable Builders being stashed in AutoValues.

The options are to exempt AV from this check, or force anyone doing this to override equals/hashCode manually (with throwing versions?). The first option seems easier.

PiperOrigin-RevId: 431130109